Changes from code review, combining classes and tests
