--- src/main/java/uk/gov/hmcts/reform/iacasedocumentsapi/domain/templates/bail/helper/BailEndApplicationTemplateHelper.java
@@ -1,65 +1,0 @@
-package uk.gov.hmcts.reform.iacasedocumentsapi.domain.templates.bail.helper;
-
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.BailCaseFieldDefinition.APPLICANT_FAMILY_NAME;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.BailCaseFieldDefinition.APPLICANT_GIVEN_NAMES;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.BailCaseFieldDefinition.BAIL_REFERENCE_NUMBER;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.BailCaseFieldDefinition.END_APPLICATION_DATE;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.BailCaseFieldDefinition.END_APPLICATION_OUTCOME;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.BailCaseFieldDefinition.END_APPLICATION_REASONS;
-import static uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.BailCaseFieldDefinition.HOME_OFFICE_REFERENCE_NUMBER;
-
-import com.google.common.base.Strings;
-import java.time.LocalDate;
-import java.time.format.DateTimeFormatter;
-import java.util.HashMap;
-import java.util.Map;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.stereotype.Component;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.BailCase;
-import uk.gov.hmcts.reform.iacasedocumentsapi.domain.entities.ccd.CaseDetails;
-import uk.gov.hmcts.reform.iacasedocumentsapi.infrastructure.CustomerServicesProvider;
-
-@Component
-public class BailEndApplicationTemplateHelper {
-
-    private static final DateTimeFormatter DOCUMENT_DATE_FORMAT = DateTimeFormatter.ofPattern("ddMMyyyy");
-    private final CustomerServicesProvider customerServicesProvider;
-    private final String govCallChargesUrl;
-
-    public BailEndApplicationTemplateHelper(@Value("${govCallChargesUrl}") String govCallChargesUrl,
-                                              CustomerServicesProvider customerServicesProvider) {
-        this.govCallChargesUrl = govCallChargesUrl;
-        this.customerServicesProvider = customerServicesProvider;
-    }
-
-    public Map<String, Object> getCommonMapFieldValues(
-        CaseDetails<BailCase> caseDetails
-    ) {
-        final BailCase bailCase = caseDetails.getCaseData();
-
-        final Map<String, Object> fieldValues = new HashMap<>();
-
-        fieldValues.put("hmcts", "[userImage:hmcts.png]");
-        fieldValues.put("applicantGivenNames", bailCase.read(APPLICANT_GIVEN_NAMES, String.class));
-        fieldValues.put("applicantFamilyName", bailCase.read(APPLICANT_FAMILY_NAME, String.class));
-        fieldValues.put("bailReferenceNumber", bailCase.read(BAIL_REFERENCE_NUMBER, String.class).orElse(""));
-        fieldValues.put("homeOfficeReferenceNumber", bailCase.read(HOME_OFFICE_REFERENCE_NUMBER, String.class).orElse(""));
-        fieldValues.put("endApplicationOutcome", bailCase.read(END_APPLICATION_OUTCOME, String.class).orElse(""));
-        fieldValues.put("endApplicationReasons", bailCase.read(END_APPLICATION_REASONS, String.class).orElse(""));
-        fieldValues.put("endApplicationDate", formatDateForRendering(bailCase.read(END_APPLICATION_DATE, String.class).orElse("")));
-        fieldValues.put("customerServicesTelephone", customerServicesProvider.getCustomerServicesTelephone());
-        fieldValues.put("customerServicesEmail", customerServicesProvider.getCustomerServicesEmail());
-        fieldValues.put("govCallChargesLink", govCallChargesUrl);
-
-        return fieldValues;
-    }
-
-    private String formatDateForRendering(
-        String date
-    ) {
-        if (!Strings.isNullOrEmpty(date)) {
-            return LocalDate.parse(date).format(DOCUMENT_DATE_FORMAT);
-        }
-        return "";
-    }
-}
